AI-Brain-service/requirements.txt
# --- Core ---
fastapi>=0.115.0
eval_type_backport>=0.3.0
uvicorn[standard]>=0.30.0
pydantic>=2.10.0
pydantic-settings>=2.7.0

# --- LLM Providers ---
openai>=1.60.0
anthropic>=0.42.0

# --- Embeddings (query embedding before retrieval) ---
tiktoken>=0.7.0

# --- Vector stores (read-only retrieval) ---
faiss-cpu>=1.7.0
numpy>=1.24.0

# --- Conversation memory ---
redis>=5.0.0

# --- Testing ---
pytest>=8.3.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
fakeredis>=2.20.0

# --- Optional (install separately if needed) ---
# Local embeddings (needs torch, ~2 GB):
#   pip install sentence-transformers>=3.0.0
# Pinecone vector store:
#   pip install pinecone-client>=5.0.0
